web_app/requirements.txt
Flask==2.2.2
kafka-python==2.0.2
psycopg2-binary==2.9.9
boto3==1.35.97